requirements.txt
django>=5.0.2
django-ninja>=1.1.0
bcrypt>=4.1.2
dynaconf>=3.2.4
psycopg2_binary>=2.9.9
